:wrench: improve repo (and upgrade all dependencies)
