+++ apps/rci-sandbox/index.html
@@ -11,6 +11,6 @@
 
   <body>
     <div id="root"></div>
-    <script src="index.tsx" type="module"></script>
+    <script src="/src/index.tsx" type="module"></script>
   </body>
 </html>
