+++ packages/use-is-focused/package.json
@@ -6,12 +6,11 @@
   "scripts": {
     "typecheck": "tsc",
     "clean": "rm -rf ./dist",
-    "build": "rollup --configPlugin rollup-plugin-ts --config rollup.config.ts",
-    "dev": "pnpm run build -- --watch"
+    "build": "rollup --config rollup.config.mjs",
+    "dev": "pnpm run build --watch"
   },
   "files": [
-    "dist",
-    "src"
+    "dist"
   ],
   "repository": {
     "type": "git",
@@ -21,28 +20,31 @@
   "sideEffects": false,
   "type": "module",
   "types": "dist/index.d.ts",
-  "main": "dist/cjs/index.cjs",
-  "module": "dist/esm/index.js",
+  "main": "dist/index.cjs",
+  "module": "dist/index.js",
   "exports": {
     ".": {
-      "import": "./dist/esm/index.js",
-      "require": "./dist/cjs/index.cjs",
-      "types": "./dist/index.d.ts"
-    }
+      "types": "./dist/index.d.ts",
+      "module": "./dist/index.js",
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs",
+      "default": "./dist/index.js"
+    },
+    "./package.json": "./package.json"
   },
   "peerDependencies": {
     "react": ">=16.8",
     "react-dom": ">=16.8"
   },
   "devDependencies": {
-    "@types/node": "^17.0.10",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "rollup": "^2.64.0",
-    "rollup-plugin-ts": "^2.0.5",
-    "tslib": "^2.3.1",
-    "typescript": "^4.5.4"
+    "@types/node": "^18.11.0",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "rollup": "^3.2.2",
+    "rollup-plugin-ts": "^3.0.2",
+    "tslib": "^2.4.0",
+    "typescript": "^4.8.4"
   }
 }
